requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
rawpy>=0.24.0
imagehash==4.3.1
Pillow==10.4.0
send2trash==1.8.3
pywebview==5.1
exifread==3.0.0
websockets==12.0
pyinstaller>=6.0
